Predecessor_Code/code/requirements.txt
absl-py==1.2.0
appdirs==1.4.4
audioread==2.1.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cycler==0.11.0
decorator==5.1.1
fonttools==4.34.4
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
joblib==1.1.0
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numba==0.56.0
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
torch==1.10.0+cu102
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu102
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
